users/views.py
import json
from utils.decorator import login_required
import bcrypt
import jwt
import requests
import boto3
import time

from django.views          import View
from django.http           import JsonResponse

from json.decoder          import JSONDecodeError

from my_settings           import SECRET, ALGORITHM, JWT_DURATION_SEC
from werecord.settings     import AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY

from users.models          import User, UserType, Batch

# 로그인
class GoogleLoginView(View):
    def get(self, request):
        id_token     = request.headers.get('Authorization', None)

        if not id_token:
            return JsonResponse({'message': 'ID_TOKEN_REQUIRED'}, status=401)

        user_request = requests.get(f'https://oauth2.googleapis.com/tokeninfo?id_token={id_token}')
        user_info    = user_request.json()

        if not user_info:
            return JsonResponse({'message': 'INVALID_TOKEN'}, status=401)

        login_user, created = User.objects.get_or_create(
            google_login_id   = user_info['sub'],
        )

        werecord_token = jwt.encode(
                {
                    'user_id'  : login_user.id,
                    'iat'      : int(time.time()),
                    'exp'      : int(time.time()) + JWT_DURATION_SEC
                }, SECRET['secret'],ALGORITHM
        )

        user_info = {
            'user_id'          : login_user.id,
            'user_type'        : login_user.user_type.name if login_user.user_type else "",
            'batch'            : login_user.batch.name if login_user.batch else "",
            "email'"           : user_info.get('email') if user_info.get('email') else "",
            "profile_image_url": user_info.get('picture') if user_info.get('picture') else "",
        }

        return JsonResponse({'user_info' : user_info, 'werecord_token' : werecord_token}, status=200)